degugging components still needed
