+++ src/components/dieComponents/Die.js
@@ -13,18 +13,10 @@ export default function Die(props) {
     }))
   }
 
-  const diceElements = Die.map(die => (
-    <Die
-      key={die.id}
-      value ={value.id}
-      isHeld={die.isHeld}
-      holdDice={() => holdDice(die.id)}
-    />
-  ))
 
   return (
     <div className='dice--face'>
-      <h2 className='die--num'>{diceElements}</h2>
+      <h2 className='die--num'>{props}</h2>
     </div>
   )
 }
